Move Analytics component back to the root dir
