+++ react/index.tsx
@@ -10,11 +10,11 @@ import { OrderGroupContext } from './components/OrderGroupContext'
 import withoutSSR from './components/WithoutSSR'
 import ErrorMessage from './components/ErrorMessage'
 import GET_ORDER_GROUP from './graphql/getOrderGroup.graphql'
-import NotFound from './Icons/NotFound'
-import Forbidden from './Icons/Forbidden'
+import NotFound from './icons/NotFound'
+import Forbidden from './icons/Forbidden'
 import Skeleton from './Skeleton'
 import { CurrencyContext } from './components/CurrencyContext'
-import { Analytics } from './components/Analytics'
+import { Analytics } from './Analytics'
 
 interface OrderGroupData {
   orderGroup: OrderGroup
@@ -111,13 +111,16 @@ const OrderPlaced: FC = () => {
 
         <article className="pt9 sans-serif">
           <Analytics eventList={orderGroup.analyticsData} />
+
           <ExtensionPoint id="order-placed-top" orderGroup={orderGroup} />
+
           <header>
             <ExtensionPoint id="order-placed-confirmation" />
             <ExtensionPoint id="order-placed-notices" />
             <ExtensionPoint id="order-placed-summary" />
             <ExtensionPoint id="order-placed-bank-invoice" />
           </header>
+
           <div className="mv6 w-80-ns w-90 center">
             {orderGroup.orders.map((order, i, { length }) => (
               <Fragment key={order.orderId}>
